Rename service files to camelCase per LAA convention
- Rename BaseApiService.ts → baseApiService.ts
- Rename ExampleApiService.ts → exampleApiService.ts
- Rename BaseApiService.spec.ts → baseApiService.spec.ts
- Update all import paths to use camelCase filenames
- Update documentation references to reflect naming convention

Addresses review feedback: TypeScript files should use camelCase
even when containing classes, following established LAA project
conventions seen in laa-manage-your-civil-cases and Express.js
best practices.

All tests passing (73/73) ✅
