+++ docs/source/documentation/features.html.md.erb
@@ -177,7 +177,7 @@ For more complex API integrations, the template provides a `BaseApiService` abst
 
 ```typescript
 // Create your own API service
-import { BaseApiService } from '#src/services/BaseApiService.js';
+import { BaseApiService } from '#src/services/baseApiService.js';
 
 class MyApiService extends BaseApiService {
   constructor() {
