+++ tests/unit/src/services/baseApiService.spec.ts
@@ -6,7 +6,7 @@
  */
 
 import { expect } from 'chai';
-import { BaseApiService } from '#src/services/BaseApiService.js';
+import { BaseApiService } from '#src/services/baseApiService.js';
 import type { AxiosInstanceWrapper } from '#types/axios-instance-wrapper.js';
 import { create } from 'middleware-axios';
